requirements.txt
pydantic
gql
requests-toolbelt
python-dotenv